requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain
langchain-box
box-sdk-gen[JWT]
openai
chromadb
langchain-openai
tiktoken
python-dotenv
bs4
langchainhub
langgraph
jupyterlab
notebook
voila
pymongo[srv]